feat: move IdentityLocRequest to loc.
logion-network/logion-internal#1197
